[refactor] Merge taichi.misc into taichi.tools.
`taichi_logo` uses high level `ti.func` decorator so importing it can
easily cause circular import. This PR moves it out of `tools/__init__.py`.

[ghstack-poisoned]
