+++ python/taichi/lang/snode.py
@@ -8,7 +8,7 @@ import taichi.lang
 from taichi.core.util import ti_core as _ti_core
 from taichi.lang import impl
 from taichi.lang.field import Field
-from taichi.misc.util import deprecated
+from taichi.tools.util import deprecated
 
 
 class SNode:
