--- python/taichi/misc/__init__.py
@@ -1,3 +1,0 @@
-from .util import *
-
-__all__ = [s for s in dir() if not s.startswith('_')]
