+++ python/taichi/snode/fields_builder.py
@@ -3,8 +3,8 @@ from typing import Any, Optional, Sequence, Union
 from taichi.core.util import ti_core as _ti_core
 from taichi.lang import impl, snode
 from taichi.lang.exception import InvalidOperationError
-from taichi.misc.util import warning
 from taichi.snode.snode_tree import SNodeTree
+from taichi.tools.util import warning
 
 _snode_registry = _ti_core.SNodeRegistry()
