+++ python/taichi/tools/__init__.py
@@ -1,9 +1,26 @@
 from .image import imdisplay, imread, imresize, imshow, imwrite
 from .np2ply import PLYWriter
-from .patterns import taichi_logo
+from .util import *
+# Don't import taichi_logo here which will cause circular import.
+# If you need it, just import from taichi.tools.patterns
 from .video import VideoManager
 
 __all__ = [
-    'PLYWriter', 'taichi_logo', 'VideoManager', 'imdisplay', 'imread',
-    'imresize', 'imshow', 'imwrite'
+    'PLYWriter',
+    'VideoManager',
+    'imdisplay',
+    'imread',
+    'imresize',
+    'imshow',
+    'imwrite',
+    'deprecated',
+    'warning',
+    'dump_dot',
+    'dot_to_pdf',
+    'obsolete',
+    'get_kernel_stats',
+    'get_traceback',
+    'set_gdb_trigger',
+    'print_profile_info',
+    'clear_profile_info',
 ]
