+++ python/taichi/types/__init__.py
@@ -1,3 +1,3 @@
 from taichi.types.annotations import *
+from taichi.types.compound_types import CompoundType, matrix, struct, vector
 from taichi.types.primitive_types import *
-from taichi.types.compound_types import CompoundType, matrix, vector, struct
