+++ python/taichi/types/compound_types.py
@@ -1,5 +1,6 @@
 import taichi
 
+
 class CompoundType:
     pass
