+++ python/taichi/ui/gui.py
@@ -6,7 +6,7 @@ import numpy as np
 import taichi.lang
 from taichi.core import ti_core as _ti_core
 from taichi.lang.field import Field, ScalarField
-from taichi.misc.util import core_veci, deprecated
+from taichi.tools.util import core_veci, deprecated
 
 import taichi as ti
